Continue to refactor PaymentAuthWebView and related classes
- Move `PaymentAuthWebViewClient` to its own class
- Create `PaymentAuthWebChromeClient`
- Remove `PaymentAuthWebView#init`

In a follow up PR, will call `setResult` in `PaymentAuthWebViewActivity`
that correctly reflects the result of the authentication attempt (e.g.
include any encountered exceptions).
