+++ stripe/src/main/java/com/stripe/android/view/PaymentAuthWebChromeClient.kt
@@ -1,0 +1,35 @@
+package com.stripe.android.view
+
+import android.app.Activity
+import android.webkit.ConsoleMessage
+import android.webkit.JsResult
+import android.webkit.WebChromeClient
+import android.webkit.WebView
+import androidx.appcompat.app.AlertDialog
+import com.stripe.android.Logger
+import com.stripe.android.R
+
+internal class PaymentAuthWebChromeClient(
+    private val activity: Activity,
+    private val logger: Logger
+) : WebChromeClient() {
+    override fun onConsoleMessage(consoleMessage: ConsoleMessage?): Boolean {
+        consoleMessage?.message()?.let(logger::debug)
+        return super.onConsoleMessage(consoleMessage)
+    }
+
+    override fun onJsConfirm(
+        view: WebView?,
+        url: String?,
+        message: String?,
+        result: JsResult?
+    ): Boolean {
+        AlertDialog.Builder(activity, R.style.AlertDialogStyle)
+            .setMessage(message)
+            .setPositiveButton(android.R.string.ok) { _, _ -> result?.confirm() }
+            .setNegativeButton(android.R.string.cancel) { _, _ -> result?.cancel() }
+            .create()
+            .show()
+        return true
+    }
+}
